Extract Delta identity column product test to own class
It was in TestDeltaLakeAlterTableCompatibility, but that's incorrect
as the test isn't centered around ALTER TABLE.
